requirements.txt
pytorch
pyworld
pysptk
scipy
numpy
tqdm
pickle
nnmnkwii